refactor: Like 파일 구조 수정 반영, Transactional 추가
