+++ src/main/java/kr/co/imguru/domain/reply/repository/ReplySearchRepository.java
@@ -1,7 +1,7 @@
 package kr.co.imguru.domain.reply.repository;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import kr.co.imguru.domain.like.QLikeReply;
+import kr.co.imguru.domain.like.entity.QLikeReply;
 import kr.co.imguru.domain.member.entity.QMember;
 import kr.co.imguru.domain.post.entity.Post;
 import kr.co.imguru.domain.post.entity.QPost;
